+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set background color for Career Explorer in Skills Self-Evaluation [#4331](https://github.com/rokwire/illinois-app/issues/4331).
 ### Added
 - Share Skill Self-Evaluation feature [#4332](https://github.com/rokwire/illinois-app/issues/4332).
+- Added search buildings fature [#4290](https://github.com/rokwire/illinois-app/issues/4290).
 
 ## [6.0.40] - 2024-09-05
 ### Changed